Add blog posts and improve doc examples
