+++ .github/workflows/website.yml
@@ -12,10 +12,10 @@ jobs:
     - uses: actions/checkout@master
     - run: cargo install mdbook --no-default-features --features output,search
     - run: cd website && mdbook build
-    - run: echo "lettre.at" > website/_book/html/CNAME
+    - run: echo "lettre.at" > website/book/html/CNAME
     - name: Deploy
       uses: peaceiris/actions-gh-pages@v2.5.1
       env:
         ACTIONS_DEPLOY_KEY: ${{ secrets.ACTIONS_DEPLOY_KEY }}
         PUBLISH_BRANCH: gh-pages
-        PUBLISH_DIR: ./website/_book/html
+        PUBLISH_DIR: ./website/book/html
