+++ Cargo.toml
@@ -17,7 +17,7 @@ is-it-maintained-open-issues = { repository = "lettre/lettre" }
 maintenance = { status = "actively-developed" }
 
 [dependencies]
-base64 = { version = "0.11", optional = true }
+base64 = { version = "0.12", optional = true }
 bufstream = { version = "0.1", optional = true }
 # TODO to 0.5
 bytes = { version = "0.4", optional = true }
