+++ src/transport/smtp/commands.rs
@@ -9,7 +9,6 @@ use crate::{
     },
     Address,
 };
-use base64;
 use log::debug;
 use std::{
     convert::AsRef,
@@ -252,7 +251,7 @@ impl AuthCommand {
         challenge: Option<String>,
     ) -> Result<AuthCommand, Error> {
         let response = if mechanism.supports_initial_response() || challenge.is_some() {
-            Some(mechanism.response(&credentials, challenge.as_ref().map(String::as_str))?)
+            Some(mechanism.response(&credentials, challenge.as_deref())?)
         } else {
             None
         };
