+++ src/transport/smtp/error.rs
@@ -4,8 +4,6 @@ use self::Error::*;
 use crate::transport::smtp::response::{Response, Severity};
 use base64::DecodeError;
 #[cfg(feature = "native-tls")]
-use native_tls;
-use nom;
 use std::{
     error::Error as StdError,
     fmt,
