+++ website/.gitignore
@@ -1,1 +1,1 @@
-/_book
+/book
