+++ website/src/SUMMARY.md
@@ -1,6 +1,9 @@
 # Summary
 
 * [Introduction](README.md)
+* [News](posts/_index.md)
+    * [Lettre 0.10 (2020.04.12)](posts/lettre-0-10.md)
+    * [Are we email yet? (2020.04.12)](posts/are-we-email-yet.md)
 * [Creating Messages](creating-messages/email.md)
 * [Sending Messages](sending-messages/_index.md)
     * [SMTP Transport](sending-messages/smtp.md)
