+++ website/src/posts/_index.md
@@ -1,0 +1,6 @@
+# News
+
+## 2020
+
+* Apr 12 - [0.10 Release](lettre-0-10.md)
+* Apr 12 - [Are we email yet?](are-we-email-yet.md)
