+++ website/src/posts/are-we-email-yet.md
@@ -1,0 +1,13 @@
+# Are we email yet?
+
+TODO
+
+short email landscape in rust
+
+* [rust-async](https://github.com/async-email/async-smtp), provides a async SMTP transport
+  compatible with lettre 0.9 messages.
+* [Proposal for a mail WG](https://internals.rust-lang.org/t/starting-a-rust-mail-wg/11678)
+* samotop
+* mail (and other by 1aim), including async smtp implementation (with tokio 0.1). Not maintained
+  anymore, pretty complete.
+
